refactor(settings): extract shared traverseTrpcPath utility
Both useTestActionHandler and useTrpcOptionsLoaders were traversing a
dot-delimited tRPC procedure path with the same null/type-guard logic.
Extract to a single traverseTrpcPath function in src/lib/trpc-traverse.ts
so guard fixes apply in one place.
