+++ apps/pops-shell/src/app/pages/settings-page/useTestActionHandler.ts
@@ -1,14 +1,6 @@
 import { trpc } from '@/lib/trpc';
+import { traverseTrpcPath } from '@/lib/trpc-traverse';
 import { useCallback } from 'react';
-
-/**
- * Resolves a dot-delimited tRPC procedure name (e.g. "media.plex.testConnection")
- * to a callable query/mutate on the tRPC client and invokes it.
- *
- * Dynamic property traversal is unavoidable here because the procedure path
- * comes from settings manifest data at runtime. The tRPC client tree is typed
- * as a deeply nested object but we need to walk it with arbitrary string keys.
- */
 
 /** Throws if the procedure response signals a failed connection without throwing. */
 function assertConnected(result: unknown): void {
@@ -21,37 +13,27 @@ function assertConnected(result: unknown): void {
   }
 }
 
+/**
+ * Resolves a dot-delimited tRPC procedure name (e.g. "media.plex.testConnection")
+ * to a callable query/mutate on the tRPC client and invokes it.
+ *
+ * Dynamic property traversal is unavoidable here because the procedure path
+ * comes from settings manifest data at runtime.
+ */
 export function useTestActionHandler() {
   const utils = trpc.useUtils();
 
   return useCallback(
     async (procedure: string) => {
-      const parts = procedure.split('.');
-      let current: unknown = utils.client;
-      for (const part of parts) {
-        if (
-          current === null ||
-          current === undefined ||
-          (typeof current !== 'object' && typeof current !== 'function')
-        ) {
-          throw new Error(`Unknown procedure: ${procedure}`);
-        }
-        current = (current as Record<string, unknown>)[part];
-        if (current == null) throw new Error(`Unknown procedure: ${procedure}`);
-      }
-      if (current !== null && (typeof current === 'object' || typeof current === 'function')) {
-        const node = current as Record<string, unknown>;
-        if (typeof node.query === 'function') {
-          const result = await (node.query as () => Promise<unknown>)();
-          assertConnected(result);
-        } else if (typeof node.mutate === 'function') {
-          const result = await (node.mutate as (input: Record<string, never>) => Promise<unknown>)(
-            {}
-          );
-          assertConnected(result);
-        } else {
-          throw new Error(`Cannot call procedure: ${procedure}`);
-        }
+      const node = traverseTrpcPath(utils.client, procedure);
+      if (typeof node.query === 'function') {
+        const result = await (node.query as () => Promise<unknown>)();
+        assertConnected(result);
+      } else if (typeof node.mutate === 'function') {
+        const result = await (node.mutate as (input: Record<string, never>) => Promise<unknown>)(
+          {}
+        );
+        assertConnected(result);
       } else {
         throw new Error(`Cannot call procedure: ${procedure}`);
       }
